refactor(proptest): tighten _common to genuinely shared helpers
