+++ tests/proptest/op_specs/_common.py
@@ -1,8 +1,10 @@
-"""Shared types and strategies for the proptest op_specs package.
+"""Shared types for the proptest op_specs package.
 
-``OpSpec`` / ``OpSample`` are the public types; the ``_*_sample_st`` helpers
-are reused across multiple op groups (unary domain, binary broadcast,
-multi-dtype, pow variants).
+``OpSpec`` and ``OpSample`` are the public types every spec module produces;
+``_unary_sample_st`` is the only sample-strategy helper genuinely shared
+across multiple op groups (elementwise unary, activations, conv/pool input
+generation). Group-specific helpers (binary broadcast, pow variants, the
+unary-domain constants) live in their consumer module.
 """
 
 import typing as T
@@ -13,12 +15,9 @@ from hypothesis import strategies as st
 
 from torch_to_nnef.inference_target.tract import TractCheckTolerance
 
-from ...wrapper import (
-    BinaryPrimitive,
-    UnaryPrimitive,
-)
-from ..inputs import Interval, dtype_st, tensor_st
-from ..shapes import binary_broadcast_shapes_st, shape_st
+from ...wrapper import UnaryPrimitive
+from ..inputs import Interval, tensor_st
+from ..shapes import shape_st
 
 
 @dataclass(frozen=True)
@@ -44,11 +43,6 @@ class OpSpec:
     xfail_reason: T.Optional[str] = None
 
 
-# -----------------------------------------------------------------------------
-# Helpers
-# -----------------------------------------------------------------------------
-
-
 def _unary_sample_st(
     op: T.Callable[[torch.Tensor], torch.Tensor],
     domain: T.Optional[Interval],
@@ -63,146 +57,3 @@ def _unary_sample_st(
         return OpSample(inputs=(x,), kwargs={}, module=UnaryPrimitive(op))
 
     return _draw()
-
-
-def _binary_broadcast_sample_st(
-    op: T.Callable[..., torch.Tensor],
-    dtype: torch.dtype = torch.float32,
-    domain: T.Optional[Interval] = None,
-    finite: bool = True,
-) -> st.SearchStrategy[OpSample]:
-    """Build a binary-op sample strategy with mutually broadcastable shapes."""
-
-    @st.composite
-    def _draw(draw) -> OpSample:
-        sa, sb = draw(binary_broadcast_shapes_st(max_rank=4, max_dim=8))
-        a = draw(tensor_st(sa, dtype, finite=finite, domain=domain))
-        b = draw(tensor_st(sb, dtype, finite=finite, domain=domain))
-        return OpSample(inputs=(a, b), kwargs={}, module=BinaryPrimitive(op))
-
-    return _draw()
-
-
-def _binary_multi_dtype_sample_st(
-    op: T.Callable[..., torch.Tensor],
-    dtypes: T.Sequence[torch.dtype] = (torch.float32, torch.float16),
-    domain_f32: T.Optional[Interval] = None,
-    domain_f16: T.Optional[Interval] = None,
-) -> st.SearchStrategy[OpSample]:
-    """Build a binary-op sample that sweeps a list of float dtypes.
-
-    Inputs are drawn with a per-dtype domain (f16 has a tighter range to
-    keep results in its representable interval). Both inputs share the
-    drawn dtype -- broadcasting is independent.
-    """
-
-    @st.composite
-    def _draw(draw) -> OpSample:
-        dtype = draw(dtype_st(list(dtypes)))
-        domain = (
-            domain_f16
-            if dtype == torch.float16 and domain_f16 is not None
-            else domain_f32
-        )
-        sa, sb = draw(binary_broadcast_shapes_st(max_rank=4, max_dim=8))
-        a = draw(tensor_st(sa, dtype, finite=True, domain=domain))
-        b = draw(tensor_st(sb, dtype, finite=True, domain=domain))
-        return OpSample(inputs=(a, b), kwargs={}, module=BinaryPrimitive(op))
-
-    return _draw()
-
-
-def _binary_pow_int_exp_sample_st() -> st.SearchStrategy[OpSample]:
-    """Pow with integer-valued exponent tensors.
-
-    Integer exponents go through a different code path in tract (a
-    repeated-multiply or sqr/rsqr fragment for small constants -- see
-    ``torch_to_nnef/op/aten/math.py:_pow``). Cover small absolute values
-    to keep results bounded.
-    """
-
-    @st.composite
-    def _draw(draw) -> OpSample:
-        sa, sb = draw(binary_broadcast_shapes_st(max_rank=4, max_dim=6))
-        base = draw(
-            tensor_st(
-                sa,
-                torch.float32,
-                finite=True,
-                domain=Interval(0.5, 10.0),
-            )
-        )
-        # Integer-valued floats from a small range (PyTorch's pow with int
-        # exponent goes through ``aten::pow.Tensor_Tensor_int`` if the
-        # exponent is an int tensor -- here we test the float-exponent
-        # path with integer values).
-        exp_int = draw(st.integers(min_value=-3, max_value=3))
-        exponent = torch.full(sb, float(exp_int), dtype=torch.float32)
-        return OpSample(
-            inputs=(base, exponent),
-            kwargs={},
-            module=BinaryPrimitive(torch.pow),
-        )
-
-    return _draw()
-
-
-def _binary_pow_sample_st() -> st.SearchStrategy[OpSample]:
-    """Pow needs separate domains for base (>=0) and exponent (small)."""
-
-    @st.composite
-    def _draw(draw) -> OpSample:
-        sa, sb = draw(binary_broadcast_shapes_st(max_rank=4, max_dim=6))
-        base = draw(
-            tensor_st(
-                sa,
-                torch.float32,
-                finite=True,
-                domain=Interval(0.1, 100.0),
-            )
-        )
-        exponent = draw(
-            tensor_st(
-                sb,
-                torch.float32,
-                finite=True,
-                domain=Interval(-2.0, 2.0),
-            )
-        )
-        return OpSample(
-            inputs=(base, exponent),
-            kwargs={},
-            module=BinaryPrimitive(torch.pow),
-        )
-
-    return _draw()
-
-
-# -----------------------------------------------------------------------------
-# Unary float specs (15)
-# -----------------------------------------------------------------------------
-
-# Domain bounds chosen to keep outputs in a numerically meaningful range and
-# avoid trivial saturation while still exercising edge cases.
-_UNARY_TRIG_DOMAIN = Interval(-6.283, 6.283)  # ~ +/- 2*pi
-_UNARY_TAN_DOMAIN = Interval(-1.4, 1.4)  # avoid tan(pi/2) explosion
-_UNARY_EXP_DOMAIN = Interval(-30.0, 30.0)
-_UNARY_LOG_DOMAIN = Interval(1e-3, 1e4)
-_UNARY_FINITE_DOMAIN = Interval(-1e4, 1e4)
-_UNARY_SQRT_DOMAIN = Interval(0.0, 1e4)
-_UNARY_RSQRT_DOMAIN = Interval(1e-3, 1e4)
-# v1 keeps reciprocal positive-only; coverage of negatives lands in v2 once
-# the strategy supports disjoint intervals.
-_UNARY_RECIP_DOMAIN = Interval(1e-2, 1e3)
-_UNARY_TANH_DOMAIN = Interval(-30.0, 30.0)
-
-
-# Domain for inverse-trig ops (asin, acos): input must be in [-1, 1].
-_UNARY_INVTRIG_DOMAIN = Interval(-1.0, 1.0)
-# Domain for acosh: input must be in [1, inf).
-_UNARY_ACOSH_DOMAIN = Interval(1.0, 1e3)
-# Domain for atanh: input must be in (-1, 1) strict; epsilon margin avoids
-# the singularities at the boundary.
-_UNARY_ATANH_DOMAIN = Interval(-0.999, 0.999)
-# Domain for hyperbolic sinh/cosh: bounded to avoid overflow at f32.
-_UNARY_HYP_DOMAIN = Interval(-30.0, 30.0)
